refactor: move out WSEditorCommandContent to own dir
