+++ src/web/src/views/workspace/WSEditorClientConfig.tsx
@@ -22,7 +22,7 @@ import {
 import { workspaceApi, specsApi, errorHandlerApi } from "../../services";
 import DoDisturbOnRoundedIcon from "@mui/icons-material/DoDisturbOnRounded";
 import AddCircleRoundedIcon from "@mui/icons-material/AddCircleRounded";
-import { Plane, Resource } from "./WSEditorCommandContent";
+import { Plane, Resource } from "./commandContent/WSEditorCommandContent";
 import SwaggerItemSelector from "./SwaggerItemSelector";
 import AddRoundedIcon from "@mui/icons-material/AddRounded";
